+++ app/controllers/workings/service_offerings/contacts_controller.rb
@@ -4,16 +4,18 @@
 module Workings
   module ServiceOfferings
     class ContactsController < ApplicationController
+      include ControllerHelpers
+
       require_feature :working
 
       # POST /workings/1/service_offerings/1/contacts
       def create
         load_working
         load_service_offering
-        load_public_service
-        decorate_service_offering
+        decorate_service_offering!
         build_contact
 
+        # TODO: Service + listener
         if @contact.save
           Mailer.contact_email(@contact, @service_offering, @working).deliver_now
           redirect_to(
@@ -27,49 +29,8 @@ module Workings
 
       private
 
-      def load_working
-        @working ||= working_scope.find(working_params)
-      end
-
-      def load_service_offering
-        @service_offering ||= service_offering_scope.find(service_offering_params)
-      end
-
-      def load_public_service
-        @public_service ||= @service_offering.public_service
-      end
-
-      def decorate_service_offering
-        @service_offering = ServiceOfferingDecorator.(@service_offering)
-      end
-
-      def build_contact
-        @contact ||= contact_scope.build
-        @contact.attributes = contact_params
-      end
-
-      def working_params
-        params[:working_id]
-      end
-
-      def service_offering_params
-        params[:service_offering_id]
-      end
-
       def contact_params
         params[:contact] ? params.require(:contact).permit(email_or_phone: Parameters.string) : {}
-      end
-
-      def working_scope
-        Working.all
-      end
-
-      def service_offering_scope
-        ServiceOffering.friendly
-      end
-
-      def contact_scope
-        Contact.none
       end
     end
   end
